lib: Update `update_system_time` to use `u64` for seconds parameter
- Change `sec` parameter type from `u32` to `u64` in `update_system_time` for improved compatibility with Unix timestamp range.

+++ sntpc/src/utils.rs
@@ -21,7 +21,7 @@ mod windows;
 /// Args:
 /// * `sec` - Seconds since UNIX epoch start
 /// * `nsec` - Fraction of seconds from an NTP response
-pub fn update_system_time(sec: u32, nsec: u32) {
+pub fn update_system_time(sec: u64, nsec: u32) {
     let time = Utc.timestamp_opt(i64::from(sec), nsec);
 
     if let Some(time) = time.single() {
